requirements.txt
asgiref==3.3.4; python_version >= "3.6"
dj-database-url==0.5.0
dj-static==0.0.6
django-ninja==0.13.0; python_version >= "3.6"
django==3.2.3; python_version >= "3.6"
greenlet==1.1.0; python_version >= "3" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3"
gunicorn==20.1.0; python_version >= "3.5"
imdbpy==2021.4.18
lxml==4.6.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
psycopg2-binary==2.8.6; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
pydantic==1.8.2; python_full_version >= "3.6.1"
pyjwt==2.1.0; python_version >= "3.6"
python-decouple==3.4
pytz==2021.1; python_version >= "3.6"
sqlalchemy==1.4.15; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
sqlparse==0.4.1; python_version >= "3.6"
static3==0.7.0
typing-extensions==3.10.0.0; python_full_version >= "3.6.1" and python_version >= "3.6"
